# api/swagger.yaml
basePath: /api/v1
definitions:
...
  internal_api_handler.ErrorResponse:
    properties:
      code:
        type: integer
      error:
        type: string
      success:
        type: boolean
    type: object
  internal_api_handler.FileUploadResponse:
    properties:
      content_type:
        type: string
      file_id:
        type: string
      file_size:
        type: integer
      filename:
        type: string
      indexed_at:
        type: string
      metadata:
        additionalProperties: true
        type: object
      success:
        type: boolean
    type: object
...
  internal_api_handler.SuccessResponse:
    properties:
      data: {}
      message:
        type: string
      success:
        type: boolean
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: |-
    多租户文件索引服务，支持 PDF、图片、文档等文件的索引、存储和搜索
    认证使用 JWT Token，数据按租户隔离存储在独立的 OpenSearch 索引中。
  title: OpenSearch File API
  version: 1.0.0
paths:
...
  /files:
    get:
      description: 获取当前租户下的文件列表
      parameters:
      - default: 1
        description: 页码
        in: query
        name: page
        type: integer
      - default: 20
        description: 每页数量
        in: query
        name: size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_api_handler.SuccessResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_api_handler.ErrorResponse'
      security:
      - BearerAuth: []
      - X-Tenant-ID: []
      summary: 列出文件
      tags:
      - File
    post:
      consumes:
      - multipart/form-data
      description: 上传文件并建立索引，支持 PDF、图片、文档等格式
      parameters:
      - description: 文件
        in: formData
        name: file
        required: true
        type: file
      - description: 文件描述
        in: formData
        name: description
        type: string
      - collectionFormat: csv
        description: 文件标签
        in: formData
        items:
          type: string
        name: tags
        type: array
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_api_handler.FileUploadResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_api_handler.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_api_handler.ErrorResponse'
      security:
      - BearerAuth: []
      - X-Tenant-ID: []
      summary: 上传文件
      tags:
      - File
  /files/{file_id}:
    delete:
      description: 根据文件 ID 删除文件
      parameters:
      - description: 文件 ID
        in: path
        name: file_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_api_handler.SuccessResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_api_handler.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/internal_api_handler.ErrorResponse'
      security:
      - BearerAuth: []
      - X-Tenant-ID: []
      summary: 删除文件
      tags:
      - File
    get:
      description: 根据文件 ID 下载文件
      parameters:
      - description: 文件 ID
        in: path
        name: file_id
        required: true
        type: string
      produces:
      - application/octet-stream
      responses:
        "200":
          description: 文件内容
          schema:
            type: file
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_api_handler.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/internal_api_handler.ErrorResponse'
      security:
      - BearerAuth: []
      - X-Tenant-ID: []
      summary: 下载文件
      tags:
      - File
  /files/{file_id}/metadata:
    get:
      description: 根据文件 ID 获取文件元数据
      parameters:
      - description: 文件 ID
        in: path
        name: file_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_api_handler.SuccessResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_api_handler.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/internal_api_handler.ErrorResponse'
      security:
      - BearerAuth: []
      - X-Tenant-ID: []
      summary: 获取文件元数据
      tags:
      - File
...
securityDefinitions:
  BearerAuth:
    description: JWT Token，格式：Bearer <token>
    in: header
    name: Authorization
    type: apiKey
  X-Tenant-ID:
    description: 租户 ID
    in: header
    name: X-Tenant-ID
    type: apiKey
swagger: "2.0"